Move common interfaces to common service
